cbs_price_service/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
camelot-py[cv]
beautifulsoup4
requests
